Move main_window and translations modules into package directories
