+++ main.py
@@ -1,6 +1,6 @@
 import sys
 from PySide6.QtWidgets import QApplication
-from main_window import MainWindow
+from gui.main_window import MainWindow
 
 if __name__ == "__main__":
     app = QApplication(sys.argv)
